+++ changelog.md
@@ -4,7 +4,9 @@
 - Fixed Crash pressing physical keyboard keys on Android
 - Fixed Startpos Switcher not sorting startpos's by X position
 - Fixed Speedhack Triggering Safe Mode even if it is above 1.0
+- Fixed Hitbox Trail not working
 - Lowered Default Position of Startpos Switcher UI
+- Added Moveable Startpos Switcher UI
 - Rewrote Safe Mode system
 - Improved Performance by making Force Platformer only apply on level open
 - Added Main Menu Gameplay